+++ src/Gnugat/Medio/Model/Type.php
@@ -1,7 +1,19 @@
 <?php
 
+/*
+ * This file is part of the Medio project.
+ *
+ * (c) Loïc Chardonnet <loic.chardonnet@gmail.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
 namespace Gnugat\Medio\Model;
 
+/**
+ * @api
+ */
 class Type
 {
     /**
@@ -11,6 +23,8 @@ class Type
 
     /**
      * @param string $name
+     *
+     * @api
      */
     public function __construct($name)
     {
@@ -42,5 +56,15 @@ class Type
     private function getNonObjectTypes()
     {
         return array('string', 'bool', 'int', 'double', 'callable', 'resource', 'array', 'null', 'mixed');
+    }
+
+    /**
+     * @return bool
+     */
+    public function hasTypeHint()
+    {
+        $isCallableFromPhp54 = ('callable' === $this->name && version_compare(PHP_VERSION, '5.4.0') >= 0);
+
+        return ($isCallableFromPhp54 || $this->isObject() || 'array' === $this->name);
     }
 }